Split into separate test classes
Run single test with, e.g.:
mvn -Dtest=FCSToCQPConverterTest test
