+++ src/test/java/se/gu/spraakbanken/fcs/endpoint/korp/KorpEndpointSearchEngineTest.java
@@ -1,68 +1,38 @@
 package se.gu.spraakbanken.fcs.endpoint.korp;
 
 import java.io.File;
-import java.io.IOException;
 import java.io.StringWriter;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.HashMap;
-import java.util.ArrayList;
 import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 
-import javax.servlet.ServletContext;
 import javax.servlet.ServletException;
 import javax.xml.stream.XMLOutputFactory;
 import javax.xml.stream.XMLStreamWriter;
 import javax.xml.stream.XMLStreamException;
 
-import eu.clarin.sru.server.CQLQueryParser;
 import eu.clarin.sru.server.SRUConfigException;
-import eu.clarin.sru.server.SRUConstants;
 import eu.clarin.sru.server.SRUDiagnosticList;
 import eu.clarin.sru.server.SRUException;
 import eu.clarin.sru.server.SRUQueryParserRegistry;
-import eu.clarin.sru.server.SRURequest;
-import eu.clarin.sru.server.SRUScanResultSet;
-import eu.clarin.sru.server.SRUSearchResultSet;
 import eu.clarin.sru.server.SRUServerConfig;
-import eu.clarin.sru.server.SRUVersion;
 import eu.clarin.sru.server.utils.SRUServerServlet;
-import eu.clarin.sru.server.fcs.DataView;
-import eu.clarin.sru.server.fcs.EndpointDescription;
-import eu.clarin.sru.server.fcs.Layer;
-import eu.clarin.sru.server.fcs.ResourceInfo;
 import eu.clarin.sru.server.fcs.FCSQueryParser;
-import eu.clarin.sru.server.fcs.SimpleEndpointSearchEngineBase;
-import eu.clarin.sru.server.fcs.parser.Expression;
-import eu.clarin.sru.server.fcs.parser.Operator;
-import eu.clarin.sru.server.fcs.parser.QueryNode;
-import eu.clarin.sru.server.fcs.parser.QuerySegment;
-import eu.clarin.sru.server.fcs.utils.SimpleEndpointDescription;
-import eu.clarin.sru.server.fcs.utils.SimpleEndpointDescriptionParser;
 
-import org.junit.After;
 import org.junit.AfterClass;
-import org.junit.Before;
 import org.junit.BeforeClass;
-import org.junit.Ignore;
 import org.junit.Test;
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
 
 import org.eclipse.jetty.servlet.ServletHolder;
 import org.eclipse.jetty.testing.ServletTester;
 
-import se.gu.spraakbanken.fcs.endpoint.korp.cqp.FCSToCQPConverter;
 import se.gu.spraakbanken.fcs.endpoint.korp.data.json.pojo.info.*;
 import se.gu.spraakbanken.fcs.endpoint.korp.data.json.pojo.query.Query;
 
 public class KorpEndpointSearchEngineTest {
-    private final static String ENDPOINT_DESCRIPTION_PATH = "target/test-classes/se/gu/spraakbanken/fcs/endpoint/korp/korp-endpoint-description-test.xml";
     private final static String SRU_SERVER_CONFIG_PATH = "src/main/webapp/WEB-INF/sru-server-config.xml";
 
     private static Map<String, String> params = Collections.unmodifiableMap(new HashMap<String, String>() {
@@ -75,21 +45,12 @@ public class KorpEndpointSearchEngineTest {
         }
     });
     private static SRUServerConfig config;
-    private static EndpointDescription sed;
 
     private static ServletTester tester;
     private static ServletHolder holder;
 
     @BeforeClass
-    public static void parseEndpointDescription() throws SRUConfigException, ServletException {
-        try {
-            sed = SimpleEndpointDescriptionParser.parse(new File(ENDPOINT_DESCRIPTION_PATH).toURI().toURL());
-            assertEquals("http://clarin.eu/fcs/capability/basic-search", sed.getCapabilities().get(0).toString());
-            assertEquals("http://clarin.eu/fcs/capability/advanced-search", sed.getCapabilities().get(1).toString());
-        } catch (MalformedURLException mue) {
-            throw new SRUConfigException("Malformed URL");
-        }
-
+    public static void setup() throws SRUConfigException, ServletException {
         tester = new ServletTester();
         // tester.setContextPath("/");
         tester.setContextPath("http://localhost:8082/sru-server");
@@ -97,6 +58,7 @@ public class KorpEndpointSearchEngineTest {
         tester.setClassLoader(SRUServerServlet.class.getClassLoader());
         holder = tester.addServlet(SRUServerServlet.class, "/sru");
 
+        // TODO: needs to be after `tester.start()`?
         URL url;
         try {
             // SRUServerServlet.class.getClassLoader().getResource("META-INF/sru-server-config.xml");
@@ -139,191 +101,6 @@ public class KorpEndpointSearchEngineTest {
         assertNotNull(kese.getCorporaInfo());
         assertNotNull(kese.getCorporaInfo().getTime());
         // assertNotNull(kese.getCorporaInfo().getCorpus("PAROLE"));
-    }
-
-    @Test
-    public void getCapabilitiesFromDescription() throws SRUConfigException {
-        assertEquals("http://clarin.eu/fcs/capability/basic-search", sed.getCapabilities().get(0).toString());
-        assertEquals("http://clarin.eu/fcs/capability/advanced-search", sed.getCapabilities().get(1).toString());
-    }
-
-    @Test
-    public void getDataViewsFromDescription() throws SRUConfigException {
-        System.out.println("sed.getSupportedDataViews(): " + sed.getSupportedDataViews());
-        assertEquals("hits", sed.getSupportedDataViews().get(0).getIdentifier());
-        assertEquals("SEND_BY_DEFAULT", sed.getSupportedDataViews().get(0).getDeliveryPolicy().toString());
-        assertEquals("application/x-clarin-fcs-adv+xml", sed.getSupportedDataViews().get(1).getMimeType());
-        assertEquals("application/x-cmdi+xml", sed.getSupportedDataViews().get(2).getMimeType());
-        assertEquals("NEED_TO_REQUEST", sed.getSupportedDataViews().get(2).getDeliveryPolicy().toString());
-    }
-
-    @Test
-    public void getLayersFromDescription() throws SRUConfigException {
-        System.out.println("sed.getSupportedLayers(): " + sed.getSupportedLayers());
-        assertEquals("http://spraakbanken.gu.se/ns/fcs/layer/word",
-                sed.getSupportedLayers().get(0).getResultId().toString());
-        assertEquals("lemma", sed.getSupportedLayers().get(1).getType().toString());
-    }
-
-    @Test
-    public void getResourcesFromDescription() throws SRUException {
-        List<ResourceInfo> riList = sed.getResourceList("hdl:10794/sbmoderna");
-        System.out.println("Resource(0).Title: " + riList.get(0).getTitle());
-        assertEquals("hits", riList.get(0).getAvailableDataViews().get(0).getIdentifier());
-        assertEquals("SEND_BY_DEFAULT", riList.get(0).getAvailableDataViews().get(0).getDeliveryPolicy().toString());
-        assertEquals("application/x-clarin-fcs-hits+xml", riList.get(0).getAvailableDataViews().get(0).getMimeType());
-        assertEquals("https://spraakbanken.gu.se/resurser/suc", riList.get(0).getLandingPageURI());
-        assertTrue(riList.get(0).hasAvailableLayers());
-        assertEquals("word", riList.get(0).getAvailableLayers().get(0).getId());
-        assertEquals("text", riList.get(0).getAvailableLayers().get(0).getType());
-        assertNull(riList.get(0).getAvailableLayers().get(0).getQualifier());
-        assertEquals("swe", riList.get(0).getLanguages().get(0));
-        assertFalse(riList.get(0).hasSubResources());
-    }
-
-    @Test
-    public void convertCQL() throws SRUException {
-        Map<String, String> params = new HashMap<String, String>();
-        // params.put("query", "text = 'användning' AND text = 'begränsad'");
-        final String query = "text = 'användningen är begränsad'";
-        final String res = "[word = 'användningen'][word = 'är'][word = 'begränsad']";
-        params.put("query", query);
-        SRUDiagnosticList diagnostics = new Diagnostic();
-        final String resActual = FCSToCQPConverter
-                .makeCQPFromCQL((new CQLQueryParser()).parseQuery(SRUVersion.VERSION_2_0, params, diagnostics));
-
-        System.out.println(resActual);
-        assertEquals(res, resActual);
-    }
-
-    @Test
-    public void convertFCSSimple() throws SRUException {
-        Map<String, String> params = new HashMap<String, String>();
-        final String query = "[word = 'användning' & pos = 'NOUN']";
-        final String res = "[word = 'användning' & pos = 'NN']";
-        params.put("query", query);
-        // params.put("query", "[text = 'användning']");
-        SRUDiagnosticList diagnostics = new Diagnostic();
-        final String resActual = FCSToCQPConverter
-                .makeCQPFromFCS((new FCSQueryParser()).parseQuery(SRUVersion.VERSION_2_0, params, diagnostics));
-
-        System.out.println(resActual);
-        assertEquals(res, resActual);
-    }
-
-    @Test
-    public void convertFCSLemma() throws SRUException {
-        Map<String, String> params = new HashMap<String, String>();
-        final String query = "[lemma = 'Kabul' & pos = 'PROPN']";
-        final String res = "[lemma contains 'Kabul' & pos = 'PM']";
-        params.put("query", query);
-        SRUDiagnosticList diagnostics = new Diagnostic();
-        final String resActual = FCSToCQPConverter
-                .makeCQPFromFCS((new FCSQueryParser()).parseQuery(SRUVersion.VERSION_2_0, params, diagnostics));
-
-        System.out.println(resActual);
-        assertEquals(res, resActual);
-    }
-
-    @Test
-    public void convertFCSNot() throws SRUException {
-        Map<String, String> params = new HashMap<String, String>();
-        final String query = "[word = 'användning' & pos != 'NOUN']";
-        final String res = "[word = 'användning' & pos != 'NN']";
-        params.put("query", query);
-        // params.put("query", "[text = 'användning']");
-        SRUDiagnosticList diagnostics = new Diagnostic();
-        final String resActual = FCSToCQPConverter
-                .makeCQPFromFCS((new FCSQueryParser()).parseQuery(SRUVersion.VERSION_2_0, params, diagnostics));
-
-        System.out.println(resActual);
-        assertEquals(res, resActual);
-    }
-
-    @Test
-    public void convertFCSWildcard() throws SRUException {
-        Map<String, String> params = new HashMap<String, String>();
-        final String query = "[word = 'användning' & pos = 'NOUN'] [] [pos = 'ADJ'] ";
-        final String res = "[word = 'användning' & pos = 'NN'] [] [pos = '(JJ|PC|RO)'] ";
-        params.put("query", query);
-        SRUDiagnosticList diagnostics = new Diagnostic();
-        final String resActual = FCSToCQPConverter
-                .makeCQPFromFCS((new FCSQueryParser()).parseQuery(SRUVersion.VERSION_2_0, params, diagnostics));
-
-        System.out.println(resActual);
-        assertEquals(res, resActual);
-    }
-
-    @Test
-    public void convertFCSRegexCaseInsensitive() throws SRUException {
-        Map<String, String> params = new HashMap<String, String>();
-        final String query = "[word = 'användning'/c & pos = 'NOUN'] ";
-        final String res = "[word = 'användning' %c & pos = 'NN'] ";
-        params.put("query", query);
-        SRUDiagnosticList diagnostics = new Diagnostic();
-        final String resActual = FCSToCQPConverter
-                .makeCQPFromFCS((new FCSQueryParser()).parseQuery(SRUVersion.VERSION_2_0, params, diagnostics));
-
-        System.out.println(resActual);
-        assertEquals(res.trim(), resActual.trim());
-    }
-
-    @Test
-    public void convertFCSRegexIgnoreDiacritics() throws SRUException {
-        Map<String, String> params = new HashMap<String, String>();
-        final String query = "[word = 'användning'/d & pos = 'NOUN'] ";
-        final String res = "[word = 'användning' %d & pos = 'NN'] ";
-        params.put("query", query);
-        SRUDiagnosticList diagnostics = new Diagnostic();
-        final String resActual = FCSToCQPConverter
-                .makeCQPFromFCS((new FCSQueryParser()).parseQuery(SRUVersion.VERSION_2_0, params, diagnostics));
-
-        System.out.println(resActual);
-        assertEquals(res.trim(), resActual.trim());
-    }
-
-    @Test
-    @Ignore
-    public void convertFCSRegexLiteral() throws SRUException {
-        Map<String, String> params = new HashMap<String, String>();
-        final String query = "[word = '?'/l & pos = 'PUNCT'] ";
-        final String res = "[word = '?' %l & pos = '(MAD|MID|PAD)'] ";
-        params.put("query", query);
-        SRUDiagnosticList diagnostics = new Diagnostic();
-        final String resActual = FCSToCQPConverter
-                .makeCQPFromFCS((new FCSQueryParser()).parseQuery(SRUVersion.VERSION_2_0, params, diagnostics));
-
-        System.out.println(resActual);
-        // This fails right now since you get d too!
-        assertEquals(res.trim(), resActual.trim());
-    }
-
-    @Test
-    public void convertFCSOccurs() throws SRUException {
-        Map<String, String> params = new HashMap<String, String>();
-        final String query = "[word = 'användning' & pos = 'NOUN'] []{1,3} [pos = 'ADJ'] ";
-        final String res = "[word = 'användning' & pos = 'NN'] []{1,3} [pos = '(JJ|PC|RO)'] ";
-        params.put("query", query);
-        SRUDiagnosticList diagnostics = new Diagnostic();
-        final String resActual = FCSToCQPConverter
-                .makeCQPFromFCS((new FCSQueryParser()).parseQuery(SRUVersion.VERSION_2_0, params, diagnostics));
-
-        System.out.println(resActual);
-        assertEquals(res, resActual);
-    }
-
-    @Test
-    public void convertFCSOccursExact() throws SRUException {
-        Map<String, String> params = new HashMap<String, String>();
-        final String query = "[word = 'användning' & pos = 'NOUN'] []{3} [pos = 'ADJ'] ";
-        final String res = "[word = 'användning' & pos = 'NN'] []{3} [pos = '(JJ|PC|RO)'] ";
-        params.put("query", query);
-        SRUDiagnosticList diagnostics = new Diagnostic();
-        final String resActual = FCSToCQPConverter
-                .makeCQPFromFCS((new FCSQueryParser()).parseQuery(SRUVersion.VERSION_2_0, params, diagnostics));
-
-        System.out.println(resActual);
-        assertEquals(res, resActual);
     }
 
     @Test
